% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coord.R
\name{plot_coord}
\alias{plot_coord}
\title{Visualization of significant differences in ERPs using gganimation}
\usage{
plot_coord(tests_rst, frames, show, point_size = 14, high_col = "firebrick",
  low_col = "forestgreen", na_col = "grey70", text = T, text_size = 3,
  text_col = "black", circle = T, nose = T, cir_nose_col = "black",
  loop = 1, ...)
}
\arguments{
\item{tests_rst}{the testing result from \code{\link{plot_fa}}.}

\item{frames}{The time point of the ERP data}

\item{show}{Which frames to show in the animation.}

\item{point_size}{The size of the points (channels) on the scalp.}

\item{high_col}{The color for "highly" significant time point (smaller p value).}

\item{low_col}{The color for "low or no significant" time point (bigger p value).}

\item{na_col}{The color for no specified channels on full 10-10 system.}

\item{text}{Logical variable. Whether to show the channel label or not.}

\item{text_size}{The text size of the channel label (when text = TRUE).}

\item{text_col}{The text color of the channel label (when text = TRUE).}

\item{circle}{Logical variable. Whether to show the scalp circle or not.}

\item{nose}{Logical variable.  Whether to show the nose or not.}

\item{cir_nose_col}{The color of the nose (when nose = TRUE).}

\item{loop}{The parameter for adjusting animations in \code{\link[animation]{ani.options}}. Whether to iterate or not (default TRUE to iterate for infinite times)}

\item{...}{Further parameters in \code{\link[animation]{ani.options}} to adjust the animation (i.e.  interval = 0.5), and  \code{\link[gganimate]{gg_animate}} to save the animation.}
}
\description{
The plot coord takes the results of \code{\link{plot_fa}} to create a animated GIF plot for visualizing significant differences in ERPs across locations on the scalp over time.
}
\examples{
data(DirectedForgetting)
time_pt <- seq(-200, 1000, 1)

test_res <- plot_fa(data = DirectedForgetting, frames = time_pt,
                    channel = 5, subject = 1, uV = 6:1206, test = 4,
                    mode = "test_signal",
                    design = (~Subject + Condition), design0 = (~Subject), nbf = 5)

plot_coord(tests_rst = test_res$Test_Rst,
           frames = time_pt,
           show = seq(200, 500, by = 1),
           loop = 1,
           interval = 0.1,
           filename = "GIFfile.gif")
}
\author{
Chi-Lin Yu <psychilinyu@gmail.com>, Ching-Fan Sheu <csheu@mail.ncku.edu.tw>
}
\seealso{
\code{\link{plot_tete}} \code{\link{plot_fa}}
}

