% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fa.R
\name{plot_fa}
\alias{plot_fa}
\title{Visualization of the results using adaptive factor-adjustement for multiple testing of ERP data}
\usage{
plot_fa(data, frames = NULL, uV, subject = NULL, channel = NULL,
  test = NULL, mode = c("test", "test_signal"), curve.col = NULL,
  labs = list(x = "Time (ms)", y = "Amplitude (microvolt)"), ggtheme = NULL,
  order = F, scalp = FALSE, coord.mat = NULL, ylim = c(-20, 20),
  significant.col = "darkgoldenrod", significant.alpha = 0.1,
  design = NULL, design0 = NULL, ...)
}
\arguments{
\item{data}{a data frame with ERP data}

\item{frames}{The time point of the ERP data}

\item{uV}{The corresponding column indices within the input data frame for the ERP amplitudes}

\item{subject}{The corresponding column index for the subject variables (factor)}

\item{channel}{The corresponding column index for the channel variables (factor)}

\item{test}{The corresponding column index for the variables you want to compare. It could be a factor (i.e, Condition) or numeric variable (i.e, Score).}

\item{mode}{The options ("test","test_signal") of the functions control the kind of plots to be made.}

\item{curve.col}{The color for the curves. Numbers must match. (e.g. two conditions need two colors)}

\item{labs}{The labs of the plot}

\item{ggtheme}{The theme setting for ggplot2}

\item{order}{Reordering the plots by the similarities between the curves using TSclust package.}

\item{scalp}{Logical variable. Plot the curve on the scalp location (follow the coord.mat option).}

\item{coord.mat}{Read users' own coordinate matrix. Default is the full 10/10 system.}

\item{ylim}{The limits of the plot on Y axis. The setting is same for ggplot2.}

\item{significant.col}{The color used to indicated the significant time point.}

\item{significant.alpha}{The alpha of the significant.col that used to indicated the significant time point.}

\item{design}{Design matrix of the full model for the relationship between the ERP and the experimental variables. Typically the output of the function model.matrix (see the R package Causeur, David, Ching-Fan Sheu, and Mei-Chen Chu. "Significance analysis of ERP data." (2014).)}

\item{design0}{Design matrix of the null model. Typically a submodel of the full model, obtained by removing columns from design. Default is NULL, corresponding to the model with no covariates. (see the R package Causeur, David, Ching-Fan Sheu, and Mei-Chen Chu. "Significance analysis of ERP data." (2014).)}

\item{...}{Other parameters in \code{\link[ERP]{erpfatest}}.}
}
\value{
Plot A "ggplot2" plot will automatically generate.

Test_Rst A same testing results of \code{\link[ERP]{erpfatest}}.
}
\description{
the function plot fa performs factor-adjusted multiple testing of ERP curves in the linear model framework (Causeur et al, 2012) and display results the same way as that of \code{\link{plot_tete}}.
}
\examples{
data(DirectedForgetting)
time_pt <- seq(-200, 1000, 1)

dta_c <- DirectedForgetting \%>\%
         filter(Channel \%in\% c("FZ","CZ","PZ")) \%>\%
         droplevels()

plot_fa(data = dta_c, 
        frames = time_pt,
        channel = 5, subject = 1, uV = 6:1206, test = 4,
        mode = "test_signal",
        design = (~Subject + Condition), design0 = (~Subject), nbf = 5)
}
\seealso{
\code{\link{plot_tete}} \code{\link{plot_coord}}  \code{\link[ERP]{erpfatest}}
}
\author{
Chi-Lin Yu <psychilinyu@gmail.com>, Ching-Fan Sheu <csheu@mail.ncku.edu.tw>
}
