% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tete.R
\name{plot_tete}
\alias{plot_tete}
\title{Routine Visualization of ERP data}
\usage{
plot_tete(data, frames = NULL, uV, subject = NULL, channel = NULL,
  test = NULL, mode = c("raw", "mean", "bootci"), order = F,
  curve.col = NULL, labs = list(x = "Time (ms)", y =
  "Amplitude (microvolt)"), ggtheme = NULL, scalp = FALSE,
  coord.mat = NULL, ylim = c(-20, 20), curve.fun = function(x, d) {    
  return(mean(x[d])) }, boot.num = 200, boot.intval = c(0.025, 0.975),
  ci.alpha = 0.3)
}
\arguments{
\item{data}{a data frame with ERP data}

\item{frames}{The time point of the ERP data}

\item{uV}{The corresponding column indices within the input data frame for the ERP amplitudes}

\item{subject}{The corresponding column index for the subject variables (factor)}

\item{channel}{The corresponding column index for the channel variables (factor)}

\item{test}{The corresponding column index for the variables you want to compare. It could be a factor (i.e, Condition) or numeric variable (i.e, Score).}

\item{mode}{The options ("raw","mean","bootci") of the functions control the kind of plots to be made.}

\item{order}{Reordering the plots by the similarities between the curves using TSclust package.}

\item{curve.col}{The color for the curves. Numbers must match. (e.g. two conditions need two colors)}

\item{labs}{The labs of the plot}

\item{ggtheme}{The theme setting for ggplot2}

\item{scalp}{Logical variable. Plot the curve on the scalp location (follow the coord.mat option).}

\item{coord.mat}{Read users' own coordinate matrix. Default is the full 10/10 system.}

\item{ylim}{The limits of the plot on Y axis. The setting is same for ggplot2.}

\item{curve.fun}{The function you want to aggreagate when model == mean or bootci. For example, curve.fun = function(x, d){return(mean(x[d]))}. See \code{\link[boot]{boot}} for further information.}

\item{boot.num}{The number of bootstrap replicates (when mode == bootci). Usually this will be a single positive integer (Default is 200).}

\item{boot.intval}{A scalar or vector containing the confidence levels (when mode == bootci), i.e, boot.intval = c(0.025,0.975).}

\item{ci.alpha}{The alpha value (the ggplot2 setting) for the ribbon area of the confidence interval (when mode == bootci).}
}
\description{
This fucntion generates routine visualization of individual or summarized ERP curves for selected channels conditioned on a factor variable with or without a grouping variable.
}
\examples{
data(DirectedForgetting)
time_pt <- seq(-200, 1000, 1)

plot_tete(data = DirectedForgetting, 
          frames = time_pt, 
          channel = 5, subject = 1, uV = 6:1206, test = 4,
          mode = "mean", scalp = TRUE)

dta_c <- filter(DirectedForgetting, 
                Channel \%in\% c("FZ", "FCZ", "CZ", "PZ")) 
                \%>\% droplevels()
plot_tete(data = dta_c, 
          frmaes = time_pt, 
          channel = 5, subject = 1, uV = 6:1206, test = 4,
          mode = "raw")

dta_c <- filter(DirectedForgetting, 
                Channel == "CZ" ) 
                \%>\% droplevels()
plot_tete(data = dta_c, 
          frames = time_pt, 
          channel = 5, subject = 1, uV = 6:1206, test = 4,
          mode = "bootci")
}
\seealso{
\code{\link{plot_fa}} \code{\link{plot_coord}}
}
\author{
Chi-Lin Yu <psychilinyu@gmail.com>, Ching-Fan Sheu <csheu@mail.ncku.edu.tw>
}
